+++ src/core/message.h
@@ -46,7 +46,7 @@ extern uint32_t nni_msg_trim_u32(nni_msg *);
 // It's faster than trim and append, but logically equivalent.
 extern uint32_t nni_msg_header_peek_u32(nni_msg *);
 extern void     nni_msg_header_poke_u32(nni_msg *, uint32_t);
-extern void     nni_msg_set_pipe(nni_msg *, uint32_t);
+extern void     nni_msg_set_pipe(nni_msg *, uint64_t);
 extern uint64_t nni_msg_get_pipe(const nni_msg *);
 
 // Reference counting messages. This allows the same message to be
